# docs/schema/V1/openapi.v1.serviceowner.verified.json
{
  "components": {
    "schemas": {
...
      "CreateDialog": {
        "additionalProperties": false,
        "properties": {
          "activities": {
            "description": "An immutable list of activities associated with the dialog.",
            "items": {
              "$ref": "#/components/schemas/CreateDialogActivity"
            },
            "nullable": true,
            "type": "array"
          },
          "apiActions": {
            "description": "The API actions associated with the dialog. Should be used in specialized, non-browser-based integrations.",
            "items": {
              "$ref": "#/components/schemas/CreateDialogApiAction"
            },
            "nullable": true,
            "type": "array"
          },
          "attachments": {
            "description": "The attachments associated with the dialog (on an aggregate level).",
            "items": {
              "$ref": "#/components/schemas/CreateDialogAttachment"
            },
            "nullable": true,
            "type": "array"
          },
          "content": {
            "description": "The dialog unstructured text content.",
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CreateDialogContent"
              }
            ]
          },
          "createdAt": {
            "description": "If set, will override the date and time when the dialog is set as created.\nIf not supplied, the current date /time will be used.",
            "example": "2022-12-31T23:59:59Z",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "dueAt": {
            "description": "The due date for the dialog. Dialogs past due date might be marked as such in frontends but will still be available.",
            "example": "2022-12-31T23:59:59Z",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "expiresAt": {
            "description": "The expiration date for the dialog. This is the last date when the dialog is available for the end user.\n            \nAfter this date is passed, the dialog will be considered expired and no longer available for the end user in any\nAPI. If not supplied, the dialog will be considered to never expire. This field can be changed after creation.",
            "example": "2022-12-31T23:59:59Z",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "extendedStatus": {
            "description": "Arbitrary string with a service-specific indicator of status, typically used to indicate a fine-grained state of\nthe dialog to further specify the \u0022status\u0022 enum.",
            "nullable": true,
            "type": "string"
          },
          "externalReference": {
            "description": "Arbitrary string with a service-specific reference to an external system or service.",
            "nullable": true,
            "type": "string"
          },
          "guiActions": {
            "description": "The GUI actions associated with the dialog. Should be used in browser-based interactive frontends.",
            "items": {
              "$ref": "#/components/schemas/CreateDialogGuiAction"
            },
            "nullable": true,
            "type": "array"
          },
          "id": {
            "description": "A self-defined UUIDv7 may be provided to support idempotent creation of dialogs. If not provided, a new UUIDv7 will be generated.",
            "example": "01913cd5-784f-7d3b-abef-4c77b1f0972d",
            "format": "guid",
            "nullable": true,
            "type": "string"
          },
          "idempotentKey": {
            "description": "An optional key to ensure idempotency in dialog creation. If provided, it allows for the safe re-submission of the same dialog creation request without creating duplicate entries.",
            "nullable": true,
            "type": "string"
          },
          "isApiOnly": {
            "description": "Indicates if this dialog is intended for API consumption only and should not be displayed in user interfaces.\nWhen true, the dialog will not be visible in portals designed for human users, but will remain accessible via API.",
            "type": "boolean"
          },
          "party": {
            "description": "The party code representing the organization or person that the dialog belongs to in URN format.",
            "example": "urn:altinn:person:identifier-no:01125512345\nurn:altinn:organization:identifier-no:912345678",
            "type": "string"
          },
          "precedingProcess": {
            "description": "Optional preceding process identifier to indicate the business process that preceded the process indicated in the \u0022Process\u0022 field. Cannot be set without also \u0022Process\u0022 being set.",
            "nullable": true,
            "type": "string"
          },
          "process": {
            "description": "Optional process identifier used to indicate a business process this dialog belongs to.",
            "nullable": true,
            "type": "string"
          },
          "progress": {
            "description": "Advisory indicator of progress, represented as 1-100 percentage value. 100% representing a dialog that has come\nto a natural completion (successful or not).",
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "searchTags": {
            "description": "A list of words (tags) that will be used in dialog search queries. Not visible in end-user DTO.",
            "items": {
              "$ref": "#/components/schemas/CreateDialogTag"
            },
            "nullable": true,
            "type": "array"
          },
          "serviceOwnerContext": {
            "description": "Metadata about the dialog owned by the service owner.",
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CreateDialogServiceOwnerContext"
              }
            ]
          },
          "serviceResource": {
            "description": "The service identifier for the service that the dialog is related to in URN-format.\nThis corresponds to a resource in the Altinn Resource Registry, which the authenticated organization\nmust own, i.e., be listed as the \u0022competent authority\u0022 in the Resource Registry entry.",
            "example": "urn:altinn:resource:some-service-identifier",
            "type": "string"
          },
          "status": {
            "description": "The aggregated status of the dialog.",
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/DialogStatusInput"
              }
            ]
          },
          "systemLabel": {
            "description": "Set the system label of the dialog.",
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/SystemLabel"
              }
            ]
          },
          "transmissions": {
            "description": "The immutable list of transmissions associated with the dialog.",
            "items": {
              "$ref": "#/components/schemas/CreateDialogTransmission"
            },
            "nullable": true,
            "type": "array"
          },
          "updatedAt": {
            "description": "If set, will override the date and time when the dialog is set as last updated.\nIf not supplied, the current date /time will be used.",
            "example": "2022-12-31T23:59:59Z",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "visibleFrom": {
            "description": "The timestamp when the dialog should be made visible for authorized end users. If not provided, the dialog will be\nimmediately available.",
            "example": "2022-12-31T23:59:59Z",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "CreateDialogServiceOwnerContext": {
        "additionalProperties": false,
        "properties": {
          "serviceOwnerLabels": {
            "description": "A list of labels, not visible in end-user APIs.",
            "items": {
              "$ref": "#/components/schemas/CreateDialogServiceOwnerLabel"
            },
            "nullable": true,
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "CreateDialogTag": {
        "additionalProperties": false,
        "properties": {
          "value": {
            "description": "A search tag value.",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "DeletedFilter": {
        "description": "",
        "enum": [
          "Exclude",
          "Include",
          "Only"
        ],
        "type": "string",
        "x-enumNames": [
          "Exclude",
          "Include",
          "Only"
        ]
      },
...
      "DialogStatus": {
        "description": "",
        "enum": [
          "InProgress",
          "Draft",
          "RequiresAttention",
          "Completed",
          "NotApplicable",
          "Awaiting"
        ],
        "type": "string",
        "x-enumNames": [
          "InProgress",
          "Draft",
          "RequiresAttention",
          "Completed",
          "NotApplicable",
          "Awaiting"
        ]
      },
      "DialogStatusInput": {
        "description": "",
        "enum": [
          "New",
          "InProgress",
          "Draft",
          "Sent",
          "RequiresAttention",
          "Completed",
          "NotApplicable",
          "Awaiting"
        ],
        "type": "string",
        "x-enumNames": [
          "New",
          "InProgress",
          "Draft",
          "Sent",
          "RequiresAttention",
          "Completed",
          "NotApplicable",
          "Awaiting"
        ]
      },
...
      "PaginatedListOfDialogListItem": {
        "additionalProperties": false,
        "properties": {
          "continuationToken": {
            "description": "The continuation token to be used to fetch the next page of items",
            "example": "createdat_2024-07-31T09:09:03.0257090Z,id_0c089101-b7cf-a476-955c-f00a78d74a4e",
            "nullable": true,
            "type": "string"
          },
          "hasNextPage": {
            "description": "Whether there are more items available that can be fetched by supplying the continuation token",
            "type": "boolean"
          },
          "items": {
            "description": "The paginated list of items",
            "items": {
              "$ref": "#/components/schemas/DialogListItem"
            },
            "nullable": true,
            "type": "array"
          },
          "orderBy": {
            "description": "The current sorting order of the items",
            "example": "contentupdatedat_desc,id_desc",
            "type": "string"
          }
        },
        "type": "object"
      },
      "ProblemDetails": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "nullable": true,
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ProblemDetails_Error"
            },
            "nullable": true,
            "type": "array"
          },
          "instance": {
            "default": "/api/route",
            "type": "string"
          },
          "status": {
            "default": 400,
            "format": "int32",
            "type": "integer"
          },
          "title": {
            "default": "One or more validation errors occurred.",
            "type": "string"
          },
          "traceId": {
            "default": "0HMPNHL0JHL76:00000001",
            "type": "string"
          },
          "type": {
            "default": "https://www.rfc-editor.org/rfc/rfc7231#section-6.5.1",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "SystemLabel": {
        "description": "",
        "enum": [
          "Default",
          "Bin",
          "Archive",
          "MarkedAsUnopened",
          "Sent"
        ],
        "type": "string",
        "x-enumNames": [
          "Default",
          "Bin",
          "Archive",
          "MarkedAsUnopened",
          "Sent"
        ]
      },
...
    },
    "securitySchemes": {
      "JWTBearerAuth": {
        "bearerFormat": "JWT",
        "description": "Enter a JWT token to authorize the requests...",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Dialogporten API description for both enduser and serviceowner users, as well as open metadata information for public key material.\u003Cbr\u003E\u003Cbr\u003EAll operations* described within this document require authentication and authorization. Read more at \u003Ca href=\u0022https://docs.altinn.studio/en/dialogporten/user-guides/authenticating/\u0022\u003Ehttps://docs.altinn.studio/en/dialogporten/user-guides/authenticating/\u003C/a\u003E\u003Cbr\u003E\u003Cbr\u003E\u003Cstrong\u003EAll GET operations* and POST operations may return or contain, respectively, personal identifiable information (national identity numbers and names).\u003C/strong\u003E\u003Cbr\u003E\u003Cbr\u003EFor more information about this product, see \u003Ca href=\u0022https://docs.altinn.studio/en/dialogporten\u0022\u003Ehttps://docs.altinn.studio/en/dialogporten\u003C/a\u003E\u003Cbr\u003E\u003Cbr\u003E\u003Cem\u003E* Except the metadata APIs",
    "title": "Dialogporten ServiceOwner",
    "version": "v1"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/serviceowner/dialogs": {
      "get": {
        "description": "Performs a search for dialogs, returning a paginated list of dialogs.\n\n* All date parameters must contain explicit time zone. Example: 2023-10-27T10:00:00Z or 2023-10-27T10:00:00\u002B01:00\n* See \u0022continuationToken\u0022 in the response for how to get the next page of results.\n* hasNextPage will be set to true if there are more items to get.",
        "operationId": "SearchDialogs",
        "parameters": [
          {
            "description": "Filter by one or more service resources",
            "explode": true,
            "in": "query",
            "name": "serviceResource",
            "schema": {
              "items": {
                "type": "string"
              },
              "nullable": true,
              "type": "array"
            },
            "style": "form"
          },
          {
            "description": "Filter by one or more owning parties",
            "explode": true,
            "in": "query",
            "name": "party",
            "schema": {
              "items": {
                "type": "string"
              },
              "nullable": true,
              "type": "array"
            },
            "style": "form"
          },
          {
            "description": "Filter by end user id",
            "in": "query",
            "name": "endUserId",
            "schema": {
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Filter by one or more extended statuses",
            "explode": true,
            "in": "query",
            "name": "extendedStatus",
            "schema": {
              "items": {
                "type": "string"
              },
              "nullable": true,
              "type": "array"
            },
            "style": "form"
          },
          {
            "description": "Filter by external reference",
            "in": "query",
            "name": "externalReference",
            "schema": {
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Filter by status",
            "explode": true,
            "in": "query",
            "name": "status",
            "schema": {
              "items": {
                "$ref": "#/components/schemas/DialogStatus"
              },
              "nullable": true,
              "type": "array"
            },
            "style": "form"
          },
          {
            "description": "If set to \u0027include\u0027, the result will include both deleted and non-deleted dialogs. If set to \u0027exclude\u0027, the result will only include non-deleted dialogs. If set to \u0027only\u0027, the result will only include deleted dialogs",
            "in": "query",
            "name": "deleted",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/DeletedFilter"
                }
              ],
              "nullable": true
            }
          },
          {
            "description": "Only return dialogs created after this date",
            "in": "query",
            "name": "createdAfter",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Only return dialogs created before this date",
            "in": "query",
            "name": "createdBefore",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Only return dialogs updated after this date",
            "in": "query",
            "name": "updatedAfter",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Only return dialogs updated before this date",
            "in": "query",
            "name": "updatedBefore",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Only return dialogs with content updated after this date",
            "in": "query",
            "name": "contentUpdatedAfter",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Only return dialogs with content updated before this date",
            "in": "query",
            "name": "contentUpdatedBefore",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Only return dialogs that have content that has/hasn\u0027t been seen.\nIf null, no filtering is applied\nIf true, returns dialogs that have been seen\nIf false, returns dialogs that have not been seen\n            \nA dialog\u0027s content is considered seen if:\n- It has been visited by the GET .../dialogs/{dialogId} endpoint since the last content update, and\n- It does not have a system label MarkedAsUnopened.",
            "in": "query",
            "name": "isContentSeen",
            "schema": {
              "nullable": true,
              "type": "boolean"
            }
          },
          {
            "description": "Only return dialogs with due date after this date",
            "in": "query",
            "name": "dueAfter",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Only return dialogs with due date before this date",
            "in": "query",
            "name": "dueBefore",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Only return dialogs with visible-from date after this date",
            "in": "query",
            "name": "visibleAfter",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Only return dialogs with visible-from date before this date",
            "in": "query",
            "name": "visibleBefore",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Filter by process",
            "in": "query",
            "name": "process",
            "schema": {
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Filter by Display state",
            "explode": true,
            "in": "query",
            "name": "systemLabel",
            "schema": {
              "items": {
                "$ref": "#/components/schemas/SystemLabel"
              },
              "nullable": true,
              "type": "array"
            },
            "style": "form"
          },
          {
            "description": "Whether to exclude API-only dialogs from the results. Defaults to false.",
            "in": "query",
            "name": "excludeApiOnly",
            "schema": {
              "nullable": true,
              "type": "boolean"
            }
          },
          {
            "description": "Search string for free text search. Will attempt to fuzzily match in all free text fields in the aggregate",
            "in": "query",
            "name": "search",
            "schema": {
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Filter by one or more labels. Multiple labels are combined with AND, i.e., all labels must match. Supports prefix matching with \u0027*\u0027 at the end of the label. For example, \u0027label*\u0027 will match \u0027label\u0027, \u0027label1\u0027, \u0027label2\u0027, etc.",
            "explode": true,
            "in": "query",
            "name": "serviceOwnerLabels",
            "schema": {
              "items": {
                "type": "string"
              },
              "nullable": true,
              "type": "array"
            },
            "style": "form"
          },
          {
            "description": "Limit free text search to texts with this language code, e.g. \u0027nb\u0027, \u0027en\u0027. Culture codes will be normalized to neutral language codes (ISO 639). Default: search all culture codes",
            "in": "query",
            "name": "searchLanguageCode",
            "schema": {
              "nullable": true,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "orderBy",
            "schema": {
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Supply \u0022continuationToken\u0022 for the response to get the next page of results, if hasNextPage is true",
            "in": "query",
            "name": "continuationToken",
            "schema": {
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Limit the number of results per page (1-1000, default: 100)",
            "in": "query",
            "name": "limit",
            "schema": {
              "format": "int32",
              "nullable": true,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedListOfDialogListItem"
                }
              }
            },
            "description": "Successfully returned the dialog list."
          },
          "401": {
            "description": "Missing or invalid authentication token. Requires a Maskinporten-token with the scope \u0022digdir:dialogporten.serviceprovider.search\u0022."
          },
          "503": {
            "content": {
              "text/plain": {
                "schema": {
                  "example": "Service Unavailable",
                  "type": "string"
                }
              }
            },
            "description": "Service Unavailable, used when Dialogporten is in maintenance mode",
            "headers": {
              "Retry-After": {
                "description": "Delay before retrying the request. Datetime format RFC1123",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWTBearerAuth": [
              "digdir:dialogporten.serviceprovider",
              "digdir:dialogporten.serviceprovider.search"
            ]
          }
        ],
        "summary": "Gets a list of dialogs",
        "tags": [
          "Serviceowner"
        ]
      },
      "post": {
        "description": "The dialog is created with the given configuration.\n\nFor detailed information on validation rules, see [the source for create dialog validators](https://github.com/Altinn/dialogporten/tree/main/src/Digdir.Domain.Dialogporten.Application/Features/V1/ServiceOwner/Dialogs/Commands/Create/Validators)",
        "operationId": "CreateDialog",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDialog"
              }
            }
          },
          "required": true,
          "x-name": "dto"
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": "018bb8e5-d9d0-7434-8ec5-569a6c8e01fc",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "The UUID of the created dialog aggregate. A relative URL to the newly created activity is set in the \u0022Location\u0022 header.",
            "headers": {
              "Etag": {
                "description": "The new UUID ETag of the dialog",
                "example": "123e4567-e89b-12d3-a456-426614174000",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation error occurred. See problem details for a list of errors."
          },
          "401": {
            "description": "Missing or invalid authentication token. Requires a Maskinporten-token with the scope \u0022digdir:dialogporten.serviceprovider\u0022."
          },
          "403": {
            "description": "Unauthorized to create a dialog for the given serviceResource (not owned by authenticated organization or has additional scope requirements defined in policy)."
          },
          "409": {
            "description": "Dialog with IdempotentKey 01941821-ffca-73a1-9335-435a882be014 has already been created."
          },
          "422": {
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Domain error occurred. See problem details for a list of errors."
          },
          "503": {
            "content": {
              "text/plain": {
                "schema": {
                  "example": "Service Unavailable",
                  "type": "string"
                }
              }
            },
            "description": "Service Unavailable, used when Dialogporten is in maintenance mode",
            "headers": {
              "Retry-After": {
                "description": "Delay before retrying the request. Datetime format RFC1123",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWTBearerAuth": [
              "digdir:dialogporten.serviceprovider"
            ]
          }
        ],
        "summary": "Creates a new dialog",
        "tags": [
          "Serviceowner"
        ]
      }
    },
...
  }
}